requirements.txt
Flask==0.9
Flask-Login==0.1.3
Flask-SQLAlchemy==0.16
Jinja2==2.6
SQLAlchemy==0.7.9
Werkzeug==0.8.3
psycopg2==2.4.5
wsgiref==0.1.2
gunicorn==0.17.2
sqlalchemy-migrate==0.7.2
Whoosh==2.4.1
Flask-WhooshAlchemy==0.5a
WTForms==1.0.3
Flask-Mail==0.9.0
redis==2.8.0